# Conflicts:
#	target/test-classes/it/polimi/ingsw/galaxytrucker/Server/Model/CardEffectVisitorMockTest.class
#	target/test-classes/it/polimi/ingsw/galaxytrucker/Server/Model/FlightCardBoardTest.class

+++ src/main/java/it/polimi/ingsw/galaxytrucker/Server/VirtualClientRmi.java
@@ -32,7 +32,7 @@ public class VirtualClientRmi extends UnicastRemoteObject implements VirtualView
     /// METODI PER PRINTARE A CLIENT ///
 
     @Override
-    public void showUpdate(String nickname, Float firePower, int powerEngine, int credits, int position, boolean purpleAline, boolean brownAlien, int numberOfHuman, int numberOfEnergy) throws RemoteException {
+    public void showUpdate(String nickname, double firePower, int powerEngine, int credits, int position, boolean purpleAline, boolean brownAlien, int numberOfHuman, int numberOfEnergy) throws RemoteException {
         view.updateView(nickname,firePower,powerEngine,credits,position,purpleAline,brownAlien,numberOfHuman,numberOfEnergy);
     }
     @Override
